# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0",
    "title": "BDApps API Documentation",
    "x-logo": {
      "url": "../images/logo1.png"
    }
  },
...
  "paths": {
...
    "/otp/request": {
      "post": {
        "tags": [
          "otp/request"
        ],
        "summary": "Request OTP ",
        "description": "",
        "operationId": "otpRequest",
        "requestBody": {
          "content": {
            "application/json;charset=utf-8": {
              "schema": {
                "$ref": "#/components/schemas/type/allOf/4/x-otpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json;charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/type/allOf/4/x-responseOfOtpRequest"
                }
              }
            }
          }
        },
        "x-code-samples": [
          {
            "lang": "Java",
            "source": "coming soon"
          },
          {
            "lang": "PHP",
            "source": "coming soon"
          },
          {
            "lang": "javaScript",
            "source": "coming soon"
          },
          {
            "lang": ".net",
            "source": "coming soon"
          }
        ]
      }
    },
    "/otp/verify": {
      "post": {
        "tags": [
          "otp/verify"
        ],
        "summary": "Verify OTP",
        "description": "",
        "operationId": "otpVerify",
        "requestBody": {
          "content": {
            "application/json;charset=utf-8": {
              "schema": {
                "$ref": "#/components/schemas/type/allOf/4/x-otpVerify"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json;charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/type/allOf/4/x-responseOfOtpVerify"
                }
              }
            }
          }
        },
        "x-code-samples": [
          {
            "lang": "Java",
            "source": "coming soon"
          },
          {
            "lang": "PHP",
            "source": "coming soon"
          },
          {
            "lang": "javaScript",
            "source": "coming soon"
          },
          {
            "lang": ".net",
            "source": "coming soon"
          }
        ]
      }
    }
  },
  "servers": [
    {
      "url": "https://developer.bdapps.com"
    }
  ],
...
}